% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmw.R
\name{wmw}
\alias{wmw}
\title{Returns a \code{WMWS3} object.}
\usage{
wmw(x, y, alpha)
}
\arguments{
\item{x}{Vector containing data for the first label.}

\item{y}{Vector containing data for the second label.}

\item{alpha}{Significance threshold.}
}
\value{
\code{WMWS3} object.
}
\description{
Returns a \code{WMWS3} object.
}
\details{
Computes the Wilcoxon-Mann-Whitney statistic and returns a
\code{WMWS3} object for two numeric vectors.
}
