% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmw.R
\name{WMWS3}
\alias{WMWS3}
\title{WMWS3 object}
\usage{
WMWS3(stat, reject)
}
\arguments{
\item{stat}{Value of the Wilcoxon-Mann-Whitney (WMW) statistic.}

\item{reject}{Indicates if the null hypothesis for the WMW test had been rejected.}
}
\value{
Returns an object of type \code{WMWS3}.
}
\description{
Constructor for \code{WMWS3} S3 object.
}
