% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test.R
\name{wilcoxon_test}
\alias{wilcoxon_test}
\title{Wilcoxon Signed Rank and Rank-Sum Tests}
\usage{
wilcoxon_test(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  exact = NULL,
  correct = TRUE
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Performs a Wilcoxon Signed Rank test on a vector of data or two paired
vectors of data, or performs a Wilcoxon Rank-Sum test on two unpaired
vectors of data.
}
\examples{
add(1, 1)
add(10, 1)
}
